[refactor] 프로필 수정 페이지 컴포넌트 분리 및 라우트 구조 변경
