+++ apps/web/app/profile/edit/page.tsx
@@ -1,0 +1,9 @@
+import ProfileEditForm from '@/components/common/profile/edit/form-edit';
+
+export default function ProfileEditPage() {
+  return (
+    <div className="mx-auto flex min-h-screen w-full max-w-md flex-col bg-white">
+      <ProfileEditForm />
+    </div>
+  );
+}
