+++ apps/web/src/components/common/profile/edit/form-edit.tsx
@@ -1,0 +1,20 @@
+'use client';
+
+import { Button } from '@/components/ui/button';
+
+import ImageUploader from './image-uploader';
+import NicknameInput from './nickname';
+
+export default function ProfileEditForm() {
+  return (
+    <form className="flex flex-1 flex-col">
+      <ImageUploader />
+      <NicknameInput />
+      <div className="mb-8 mt-auto px-4">
+        <Button type="submit" fullWidth>
+          수정 완료
+        </Button>
+      </div>
+    </form>
+  );
+}
